API_LLM/requirements.txt
fastapi
uvicorn
ollama
python-dotenv
requests
pandas
json